Updated Tree Model for Increased Flexibility
Added tests and mock data to test this increased flexibility. Allows Directories
to be created without a length, computed size when requested and renamed recurse
to unmarshall. Added two methods to help unpack a serialized file tree.
Increased preformance of some parts of the directory flattening and added more
functionality.