Added Comparable Filters
Added filters which can be used to filter using the following syntax:

    size > 1gb
    size < 10gb

this works for downloadSpeed and uploadSpeed using the keywords download and
upload. When using multiple filters, records which pass all of the filters are
the only ones shown. I've added tests and also made sure that the filter
persists across navigation.