Added Sorting!
You can now sort by clicking on the column table headers. Clicking once, causes
the torrents to be sorted by the column in descending order, clicking again,
causes the torrents to be sorted in an ascending order. Clicking on another
column causes the torrents to be sorted by that column in descending order and
to reset the sorting, simply click on the arrow on the left which displays the
sort order. Sorting also works when filtering. Reseting the sort order when
filtering causes the torrents to be sorted by their match strength, descending.